Works Poorly
The model seems to train as expected, but it fails to learn the expected behavior.
This could be due to an error in my calculations, or it could be due to issues in the noise function I am using (random gaussian). I plan to update this once I have a fix, and will add more comments explaining this model.