Pendulum Working
The pendulum environment now works as expected! I made two basic mistakes which causes the issues before. The first were forgotten parenthesis on the 'step' function for gradient descent. This stopped the model learning. The second was that I forgot to set the next state to equal the current state at the end of the 'testing' loop.

Going forward, I want to enable more configuration of the models via command line arguments. I also want to provide more graphs, specifically of reward over time for the 'testing' mode.